add 64  bit support initially

+++ drivers/uart/uart.c
@@ -7,7 +7,7 @@
 #include <loongson/addrspace.h>
 
 #define CONS_BAUD B115200
-u8* UART_BASE = (unsigned long)PHY_TO_UNCACHED((unsigned long)UART_ADDR);
+u8* UART_BASE = PHY_TO_UNCACHED(UART_ADDR);
 
 void uart_out(u32 port, u8 c)
 {
